pyproject.toml
[project]
name = "livermetabolism-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "pyyaml",
    "rich"
]
